requirements.txt
PyQt5
PyQtWebEngine
pandas
pyserial
folium
requests
